Integration tests, logger levels, magicswap tweaks, chain switching improvements (#86)
* Add integration test for connect flow

* Add loggerLevel and autoInitialize properties to TDKConfig

* Fix Magicswap type (priceUSD). Improve error handling on UI

* Refactor TDKLogger

* Use LogDebug to log event payload. Allow config of different log level for prod

* Tweak MagicswapUI responsiveness

* Use debug logger in tests. Update integration tests

* Skip unnecessarily re-initializing thirdweb sdk when switching to same chain

* Tests for IdentityUI

* Properly reconnect when switching chains after socials login

* Refactor reconnect methods. Expose Reconnect(email) method

* Rename methods for clarity

+++ Assets/Treasure/TDK/Runtime/API/Magicswap.cs
@@ -45,7 +45,7 @@ namespace Treasure
         public string urlSlug;
         public string collectionId;
         public List<string> collectionTokenIds;
-        public int priceUSD; // TODO check type; big, int or double? examples are 0
+        public double priceUSD;
         public string reserve;
     }
